use component for dashboard
