+++ frontend/app/dashboard/page.tsx
@@ -1,263 +1,134 @@
 "use client";
 
 import React, { useState, useEffect } from "react";
-import { Table, TableHeader, TableBody, TableColumn, TableRow, TableCell } from "@heroui/table";
 import { useRouter } from "next/navigation";
-import { Button, Dropdown, DropdownItem, DropdownMenu, DropdownSection, DropdownTrigger } from "@heroui/react";
 import { addToast } from "@heroui/toast";
 
-import { DownloadIcon, SortIcon } from "@/components/icons";
+import JobApplicationsDashboard, { Application } from "@/components/JobApplicationsDashboard";
 import { checkAuth } from "@/utils/auth";
 
-interface Application {
-	id?: string;
-	company_name: string;
-	application_status: string;
-	received_at: string;
-	job_title: string;
-	subject: string;
-	email_from: string;
-}
-
-// Load sort key from localStorage or use default "Sort By"
-const storedSortKey =
-	typeof window !== "undefined" ? localStorage.getItem("sortKey") || "Date (Newest)" : "Date (Newest)";
-
 export default function Dashboard() {
-	const router = useRouter();
-	const [data, setData] = useState<Application[]>([]);
-	const [loading, setLoading] = useState(true);
-	const [downloading, setDownloading] = useState(false);
-	const [error, setError] = useState<string | null>(null);
-	const [sortedData, setSortedData] = useState<Application[]>([]);
-	const [selectedKeys, setSelectedKeys] = useState(new Set([storedSortKey]));
+  const router = useRouter();
+  const [data, setData] = useState<Application[]>([]);
+  const [loading, setLoading] = useState(true);
+  const [downloading, setDownloading] = useState(false);
+  const [error, setError] = useState<string | null>(null);
+  
+  const apiUrl = process.env.NEXT_PUBLIC_API_URL || "http://localhost:8000";
 
-	const apiUrl = process.env.NEXT_PUBLIC_API_URL || "http://localhost:8000";
-	const selectedValue = React.useMemo(() => Array.from(selectedKeys).join(", ").replace(/_/g, ""), [selectedKeys]);
-
-	useEffect(() => {
-		const fetchData = async () => {
-			try {
-				// Check if user is logged in
-				const isAuthenticated = await checkAuth(apiUrl);
-				if (!isAuthenticated) {
-					addToast({
-						title: "You need to be logged in to access this page.",
-						color: "warning"
-					});
-					router.push("/");
-					return;
-				}
-
-				// Fetch applicaions (if user is logged in)
-				const response = await fetch(`${apiUrl}/get-emails`, {
-					method: "GET",
-					credentials: "include" // Include cookies for session management
-				});
-
-				if (!response.ok) {
-					if (response.status === 404) {
-						setError("No applications found");
-					} else {
-						throw new Error(`HTTP error! status: ${response.status}`);
-					}
-				}
-
-				const result = await response.json();
-
-				if (result.length === 0) {
-					setError("No applications found");
-				} else {
-					setData(result);
-				}
-			} catch {
-				setError("Failed to load applications");
-			} finally {
-				setLoading(false);
-			}
-		};
-
-		fetchData();
-	}, [router]);
-
-	// Sort data based on selected key
-	useEffect(() => {
-		const sortData = () => {
-			const sorted = [...data];
-			const sortKey = Array.from(selectedKeys)[0];
-
-			switch (sortKey) {
-				case "Date (Newest)":
-					sorted.sort((a, b) => new Date(b.received_at).getTime() - new Date(a.received_at).getTime());
-					break;
-				case "Date (Oldest)":
-					sorted.sort((a, b) => new Date(a.received_at).getTime() - new Date(b.received_at).getTime());
-					break;
-				case "Company":
-					sorted.sort((a, b) => a.company_name.localeCompare(b.company_name));
-					break;
-				case "Job Title":
-					sorted.sort((a, b) => a.job_title.localeCompare(b.job_title));
-					break;
-				case "Status":
-					sorted.sort((a, b) => a.application_status.localeCompare(b.application_status));
-					break;
-				default:
-					break;
-			}
-			setSortedData(sorted);
-		};
-
-		if (data.length > 0) {
-			sortData();
-		}
-	}, [selectedKeys, data]);
-
-	// Handle sorting selection change and store it in localStorage
-	const handleSortChange = (keys: Set<string>) => {
-		const sortKey = Array.from(keys)[0];
-		localStorage.setItem("sortKey", sortKey);
-		setSelectedKeys(new Set([sortKey]));
-	};
-
-	async function downloadCsv() {
-		setDownloading(true);
+  useEffect(() => {
+    const fetchData = async () => {
 		try {
-			const response = await fetch(`${apiUrl}/process-csv`, {
-				method: "GET",
-				credentials: "include"
-			});
-
-			if (!response.ok) {
-				let description = "Something went wrong. Please try again.";
-
-				if (response.status === 429) {
-					description = "Download limit reached. Please wait before trying again.";
-				} else {
-					description = "Please try again or contact help@jobba.help if the issue persists.";
-				}
-
+			// Check if user is logged in
+			const isAuthenticated = await checkAuth(apiUrl);
+			if (!isAuthenticated) {
 				addToast({
-					title: "Failed to download CSV",
-					description,
-					color: "danger"
+					title: "You need to be logged in to access this page.",
+					color: "warning"
 				});
-
+				router.push("/");
 				return;
 			}
 
-			// Create a download link to trigger the file download
-			const blob = await response.blob();
-			const link = document.createElement("a");
-			const url = URL.createObjectURL(blob);
-			link.href = url;
-			link.download = `job_applications_${new Date().toISOString().split("T")[0]}.csv`;
-			document.body.appendChild(link);
-			link.click();
-			document.body.removeChild(link);
-			URL.revokeObjectURL(url);
+			// Fetch applicaions (if user is logged in)
+			const response = await fetch(`${apiUrl}/get-emails`, {
+				method: "GET",
+				credentials: "include" // Include cookies for session management
+			});
+
+			if (!response.ok) {
+				if (response.status === 404) {
+					setError("No applications found");
+				} else {
+					throw new Error(`HTTP error! status: ${response.status}`);
+				}
+			}
+
+			const result = await response.json();
+
+			if (result.length === 0) {
+				setError("No applications found");
+			} else {
+				setData(result);
+			}
 		} catch {
+			setError("Failed to load applications");
+		} finally {
+			setLoading(false);
+		}
+    };
+
+    fetchData();
+  }, [apiUrl, router]);
+
+  async function downloadCsv() {
+	setDownloading(true);
+	try {
+		const response = await fetch(`${apiUrl}/process-csv`, {
+			method: "GET",
+			credentials: "include"
+		});
+
+		if (!response.ok) {
+			let description = "Something went wrong. Please try again.";
+
+			if (response.status === 429) {
+				description = "Download limit reached. Please wait before trying again.";
+			} else {
+				description = "Please try again or contact help@jobba.help if the issue persists.";
+			}
+
 			addToast({
-				title: "Something went wrong",
-				description: "Please try again",
+				title: "Failed to download CSV",
+				description,
 				color: "danger"
 			});
-		} finally {
-			setDownloading(false);
+
+			return;
 		}
+
+		// Create a download link to trigger the file download
+		const blob = await response.blob();
+		const link = document.createElement("a");
+		const url = URL.createObjectURL(blob);
+		link.href = url;
+		link.download = `job_applications_${new Date().toISOString().split("T")[0]}.csv`;
+		document.body.appendChild(link);
+		link.click();
+		document.body.removeChild(link);
+		URL.revokeObjectURL(url);
+	} catch {
+		addToast({
+			title: "Something went wrong",
+			description: "Please try again",
+			color: "danger"
+		});
+	} finally {
+		setDownloading(false);
 	}
+}
 
-	return (
-		<div className="p-6">
-			<div className="flex items-center justify-between mb-4">
-				<h1 className="text-2xl font-bold">Job Applications Dashboard</h1>
-				<div className="flex gap-x-4">
-					<Dropdown>
-						<DropdownTrigger>
-							<Button
-								className="pl-3"
-								color="primary"
-								isDisabled={!data || data.length === 0}
-								startContent={<SortIcon />}
-								variant="bordered"
-							>
-								{selectedValue}
-							</Button>
-						</DropdownTrigger>
-						<DropdownMenu
-							disallowEmptySelection
-							aria-label="Single selection example"
-							selectedKeys={selectedKeys}
-							selectionMode="single"
-							variant="flat"
-							onSelectionChange={(keys) => handleSortChange(keys as Set<string>)}
-						>
-							<DropdownSection title="Sort By">
-								<DropdownItem key="Date (Newest)">Date Received (Newest First)</DropdownItem>
-								<DropdownItem key="Date (Oldest)">Date Received (Oldest First)</DropdownItem>
-								<DropdownItem key="Company">Company (A-Z)</DropdownItem>
-								<DropdownItem key="Job Title">Job Title (A-Z)</DropdownItem>
-								<DropdownItem key="Status">Application Status</DropdownItem>
-							</DropdownSection>
-						</DropdownMenu>
-					</Dropdown>
-					<Button
-						color="success"
-						isDisabled={!data || data.length === 0}
-						isLoading={downloading}
-						startContent={<DownloadIcon />}
-						onPress={downloadCsv}
-					>
-						Download CSV
-					</Button>
-				</div>
-			</div>
+  // Rest of your component remains unchanged
+  if (error) {
+    return (
+      <div className="p-6 flex flex-col items-center justify-center min-h-[50vh]">
+        <p className="text-red-600 mb-4">{error}</p>
+        <button 
+          className="px-4 py-2 bg-blue-600 text-white rounded"
+          onClick={() => window.location.reload()}
+        >
+          Retry
+        </button>
+      </div>
+    );
+  }
 
-			{loading ? (
-				<p>Loading applications...</p>
-			) : error ? (
-				<div className="text-red-500">
-					<p>{error}</p>
-					<Button color="warning" onPress={() => window.location.reload()}>
-						Try again
-					</Button>
-				</div>
-			) : (
-				<div className="overflow-x-auto bg-white shadow-md rounded-lg">
-					<Table aria-label="Applications Table">
-						<TableHeader>
-							<TableColumn>Company</TableColumn>
-							<TableColumn>Status</TableColumn>
-							<TableColumn>Received</TableColumn>
-							<TableColumn>Job Title</TableColumn>
-							<TableColumn>Subject</TableColumn>
-							<TableColumn>Sender</TableColumn>
-						</TableHeader>
-						<TableBody>
-							{sortedData.map((item) => (
-								<TableRow key={item.id || item.received_at}>
-									<TableCell>{item.company_name || "--"}</TableCell>
-									<TableCell>
-										<span
-											className={`inline-flex items-center justify-center px-2 py-1 rounded ${
-												item.application_status.toLowerCase() === "rejected"
-													? "bg-red-100 text-red-800"
-													: "bg-green-100 text-green-800"
-											}`}
-										>
-											{item.application_status || "--"}
-										</span>
-									</TableCell>
-									<TableCell>{new Date(item.received_at).toLocaleDateString() || "--"}</TableCell>
-									<TableCell>{item.job_title || "--"}</TableCell>
-									<TableCell className="max-w-[300px] truncate">{item.subject || "--"}</TableCell>
-									<TableCell>{item.email_from || "--"}</TableCell>
-								</TableRow>
-							))}
-						</TableBody>
-					</Table>
-				</div>
-			)}
-		</div>
-	);
+  return (
+    <JobApplicationsDashboard
+      data={data}
+      loading={loading}
+      downloading={downloading}
+      onDownloadCsv={downloadCsv}
+    />
+  );
 }
