+++ frontend/app/preview/dashboard/page.tsx
@@ -1,243 +1,105 @@
 "use client";
 
 import React, { useState, useEffect } from "react";
-import { Table, TableHeader, TableBody, TableColumn, TableRow, TableCell } from "@heroui/table";
 import {
-	Button,
-	Dropdown,
-	DropdownItem,
-	DropdownMenu,
-	DropdownSection,
-	DropdownTrigger,
-	Modal,
-	ModalBody,
-	ModalContent,
-	ModalFooter,
-	ModalHeader,
-	useDisclosure
+  Modal,
+  ModalBody,
+  ModalContent, 
+  ModalFooter,
+  ModalHeader,
+  useDisclosure,
+  Button
 } from "@heroui/react";
 import { useRouter } from "next/navigation";
 
-import { DownloadIcon, SortIcon } from "@/components/icons";
+import JobApplicationsDashboard, { Application } from "@/components/JobApplicationsDashboard";
 import { mockData } from "@/utils/mockData";
 
-// Load sort key from localStorage or use default "Sort By"
-const storedSortKey =
-	typeof window !== "undefined" ? localStorage.getItem("sortKey") || "Date (Newest)" : "Date (Newest)";
+export default function PreviewDashboard() {
+  const { isOpen, onOpen, onClose } = useDisclosure();
+  const [data, setData] = useState<Application[]>([]);
+  const [loading, setLoading] = useState(true);
+  const [downloading, setDownloading] = useState(false);
+  const router = useRouter();
 
-export default function Dashboard() {
-	interface Application {
-		id: string;
-		company_name: string;
-		application_status: string;
-		received_at: string;
-		job_title: string;
-		subject: string;
-		email_from: string;
-	}
+  useEffect(() => {
+    setLoading(true);
+    const dataTimeout = setTimeout(() => {
+      setData(mockData);
+      setLoading(false);
+    }, 1500);
 
-	const { isOpen, onOpen, onClose } = useDisclosure();
-	const [data, setData] = useState<Application[]>([]);
-	const [loading, setLoading] = useState(true);
-	const [downloading, setDownloading] = useState(false);
-	const [sortedData, setSortedData] = useState<Application[]>([]);
-	const [selectedKeys, setSelectedKeys] = useState(new Set([storedSortKey]));
-	const router = useRouter();
+    const openTimeout = setTimeout(() => {
+      onOpen();
+    }, 10000);
 
-	const selectedValue = React.useMemo(() => Array.from(selectedKeys).join(", ").replace(/_/g, ""), [selectedKeys]);
+    return () => {
+      clearTimeout(dataTimeout);
+      clearTimeout(openTimeout);
+    };
+  }, [onOpen]);
 
-	useEffect(() => {
-		setLoading(true);
-		const dataTimeout = setTimeout(() => {
-			setData(mockData);
-			setLoading(false);
-		}, 1500);
+  // Handle CSV download
+  async function downloadCsv() {
+    setDownloading(true);
+    // Mock CSV generation (no api call)
+    const mockCsvContent =
+      "Company,Status,Received,Job Title,Subject,Sender\n" +
+      mockData
+        .map(
+          (item) =>
+            `${item.company_name},${item.application_status},${new Date(item.received_at).toLocaleDateString()},${item.job_title},${item.subject},${item.email_from}`
+        )
+        .join("\n");
 
-		const openTimeout = setTimeout(() => {
-			onOpen();
-		}, 10000);
+    const blob = new Blob([mockCsvContent], { type: "text/csv" });
+    const link = document.createElement("a");
+    const url = URL.createObjectURL(blob);
+    link.href = url;
+    link.download = `job_applications_${new Date().toISOString().split("T")[0]}.csv`;
+    document.body.appendChild(link);
+    link.click();
+    document.body.removeChild(link);
+    URL.revokeObjectURL(url);
+    setDownloading(false);
+  }
 
-		return () => {
-			clearTimeout(dataTimeout);
-			clearTimeout(openTimeout);
-		};
-	}, []);
+  const PromoModal = (
+    <Modal backdrop="blur" isOpen={isOpen} size="xl" onClose={onClose}>
+      <ModalContent>
+        {(onClose) => (
+          <>
+            <ModalHeader className="flex flex-col gap-1">
+              Enjoying the preview? Join the waitlist!
+            </ModalHeader>
+            <ModalBody>
+              <p>
+                By joining the waitlist, you'll receive updates on new features and an invitation to
+                signup when we launch outside of beta.
+              </p>
+            </ModalBody>
+            <ModalFooter>
+              <Button color="danger" variant="light" onPress={onClose}>
+                Close
+              </Button>
+              <Button color="primary" onPress={() => router.push("/")}>
+                Sign Up Now
+              </Button>
+            </ModalFooter>
+          </>
+        )}
+      </ModalContent>
+    </Modal>
+  );
 
-	// Sort data based on selected key
-	useEffect(() => {
-		const sortData = () => {
-			const sorted = [...data];
-			const sortKey = Array.from(selectedKeys)[0];
-
-			switch (sortKey) {
-				case "Date (Newest)":
-					sorted.sort((a, b) => new Date(b.received_at).getTime() - new Date(a.received_at).getTime());
-					break;
-				case "Date (Oldest)":
-					sorted.sort((a, b) => new Date(a.received_at).getTime() - new Date(b.received_at).getTime());
-					break;
-				case "Company":
-					sorted.sort((a, b) => a.company_name.localeCompare(b.company_name));
-					break;
-				case "Job Title":
-					sorted.sort((a, b) => a.job_title.localeCompare(b.job_title));
-					break;
-				case "Status":
-					sorted.sort((a, b) => a.application_status.localeCompare(b.application_status));
-					break;
-				default:
-					break;
-			}
-			setSortedData(sorted);
-		};
-
-		if (data.length > 0) {
-			sortData();
-		}
-	}, [selectedKeys, data]);
-
-	// Handle sorting selection change and store it in localStorage
-	const handleSortChange = (keys: Set<string>) => {
-		const sortKey = Array.from(keys)[0];
-		localStorage.setItem("sortKey", sortKey);
-		setSelectedKeys(new Set([sortKey]));
-	};
-
-	// Handle CSV download
-	async function downloadCsv() {
-		setDownloading(true);
-		// Mock CSV generation (no api call)
-		const mockCsvContent =
-			"Company,Status,Received,Job Title,Subject,Sender\n" +
-			mockData
-				.map(
-					(item) =>
-						`${item.company_name},${item.application_status},${new Date(item.received_at).toLocaleDateString()},${item.job_title},${item.subject},${item.email_from}`
-				)
-				.join("\n");
-
-		const blob = new Blob([mockCsvContent], { type: "text/csv" });
-		const link = document.createElement("a");
-		const url = URL.createObjectURL(blob);
-		link.href = url;
-		link.download = `job_applications_${new Date().toISOString().split("T")[0]}.csv`;
-		document.body.appendChild(link);
-		link.click();
-		document.body.removeChild(link);
-		URL.revokeObjectURL(url);
-		setDownloading(false);
-	}
-
-	return (
-		<div className="p-6">
-			<Modal backdrop="blur" isOpen={isOpen} size="xl" onClose={onClose}>
-				<ModalContent>
-					{(onClose) => (
-						<>
-							<ModalHeader className="flex flex-col gap-1">
-								Enjoying the preview? Join the waitlist!
-							</ModalHeader>
-							<ModalBody>
-								<p>
-									By joining the waitlist, you'll receive updates on new features and an invitation to
-									signup when we launch outside of beta.
-								</p>
-							</ModalBody>
-							<ModalFooter>
-								<Button color="danger" variant="light" onPress={onClose}>
-									Close
-								</Button>
-								<Button color="primary" onPress={() => router.push("/")}>
-									Sign Up Now
-								</Button>
-							</ModalFooter>
-						</>
-					)}
-				</ModalContent>
-			</Modal>
-
-			<div className="flex items-center justify-between mb-4">
-				<h1 className="text-2xl font-bold">Job Applications Dashboard</h1>
-				<div className="flex gap-x-4">
-					<Dropdown>
-						<DropdownTrigger>
-							<Button
-								className="pl-3"
-								color="primary"
-								isDisabled={!data || data.length === 0}
-								startContent={<SortIcon />}
-								variant="bordered"
-							>
-								{selectedValue}
-							</Button>
-						</DropdownTrigger>
-						<DropdownMenu
-							disallowEmptySelection
-							aria-label="Single selection example"
-							selectedKeys={selectedKeys}
-							selectionMode="single"
-							variant="flat"
-							onSelectionChange={(keys) => handleSortChange(keys as Set<string>)}
-						>
-							<DropdownSection title="Sort By">
-								<DropdownItem key="Date (Newest)">Date Received (Newest First)</DropdownItem>
-								<DropdownItem key="Date (Oldest)">Date Received (Oldest First)</DropdownItem>
-								<DropdownItem key="Company">Company (A-Z)</DropdownItem>
-								<DropdownItem key="Job Title">Job Title (A-Z)</DropdownItem>
-								<DropdownItem key="Status">Application Status</DropdownItem>
-							</DropdownSection>
-						</DropdownMenu>
-					</Dropdown>
-					<Button
-						color="success"
-						isDisabled={!data || data.length === 0}
-						isLoading={downloading}
-						startContent={<DownloadIcon />}
-						onPress={downloadCsv}
-					>
-						Download CSV
-					</Button>
-				</div>
-			</div>
-
-			{loading ? (
-				<p>Loading applications...</p>
-			) : (
-				<div className="overflow-x-auto bg-white shadow-md rounded-lg">
-					<Table aria-label="Applications Table">
-						<TableHeader>
-							<TableColumn>Company</TableColumn>
-							<TableColumn>Status</TableColumn>
-							<TableColumn>Received</TableColumn>
-							<TableColumn>Job Title</TableColumn>
-							<TableColumn>Subject</TableColumn>
-							<TableColumn>Sender</TableColumn>
-						</TableHeader>
-						<TableBody>
-							{sortedData.map((item) => (
-								<TableRow key={item.id || item.received_at}>
-									<TableCell>{item.company_name || "--"}</TableCell>
-									<TableCell>
-										<span
-											className={`inline-flex items-center justify-center px-2 py-1 rounded ${
-												item.application_status.toLowerCase() === "rejected"
-													? "bg-red-100 text-red-800"
-													: "bg-green-100 text-green-800"
-											}`}
-										>
-											{item.application_status || "--"}
-										</span>
-									</TableCell>
-									<TableCell>{new Date(item.received_at).toLocaleDateString() || "--"}</TableCell>
-									<TableCell>{item.job_title || "--"}</TableCell>
-									<TableCell className="max-w-[300px] truncate">{item.subject || "--"}</TableCell>
-									<TableCell>{item.email_from || "--"}</TableCell>
-								</TableRow>
-							))}
-						</TableBody>
-					</Table>
-				</div>
-			)}
-		</div>
-	);
+  return (
+    <JobApplicationsDashboard
+      title="Preview Dashboard"
+      data={data}
+      loading={loading}
+      downloading={downloading}
+      onDownloadCsv={downloadCsv}
+      extraHeader={PromoModal}
+    />
+  );
 }
